Task 2.5: RFB protocol handshake (RFB 3.3/3.8 support)
Implement complete RFB protocol handshake with version negotiation,
security handshake (SecurityType::None), and initialization phase.

New components:
- rfb-protocol/src/handshake.rs: Version/security/init handlers (378 lines)
- rfb-common crate: Point and Rect types (95 lines)
- 7 handshake unit tests covering version and security negotiation

Updates:
- Fix Rect type parameters (i32/u32 instead of u16)
- Format existing code with cargo fmt

Tests: 140 passing (56 rfb-protocol + 19 rfb-pixelbuffer + 3 rfb-common + 62 doctests)

+++ rust-vnc-viewer/rfb-pixelbuffer/src/managed.rs
@@ -264,8 +264,8 @@ impl MutablePixelBuffer for ManagedPixelBuffer {
         // Calculate source rectangle
         // src_offset is the offset FROM destination TO source
         // So source = dest + src_offset
-        let src_x = (dest.x as i32 + src_offset.x) as u16;
-        let src_y = (dest.y as i32 + src_offset.y) as u16;
+        let src_x = dest.x + src_offset.x;
+        let src_y = dest.y + src_offset.y;
         let src_rect = Rect::new(src_x, src_y, dest.width, dest.height);
         self.validate_rect(src_rect)?;
